Move SystemTestSchema into DSL
