+++ java/system-test/system-test-suite/src/test/java/sleeper/systemtest/suite/StateStoreSnapshotsST.java
@@ -36,8 +36,8 @@ import java.util.stream.LongStream;
 import static java.util.stream.Collectors.toMap;
 import static org.assertj.core.api.Assertions.assertThat;
 import static sleeper.core.properties.table.TableProperty.STATESTORE_CLASSNAME;
+import static sleeper.systemtest.dsl.util.SystemTestSchema.DEFAULT_SCHEMA;
 import static sleeper.systemtest.suite.fixtures.SystemTestInstance.MAIN;
-import static sleeper.systemtest.suite.fixtures.SystemTestSchema.DEFAULT_SCHEMA;
 
 @SystemTest
 public class StateStoreSnapshotsST {
